Rename useAutoExpandTextArea
